+++ lib/golay24.c
@@ -23,15 +23,15 @@ extern "C" {
 
 #include "golay24.h"
 
-// Replacement for __builtin_parity in compilers other than GCC
-#ifndef __GNUC__
+// Replacement for __builtin_parity in compilers other than GCC and clang
+#if !defined(__GNUC__) && !defined(__clang__)
 static inline uint32_t __builtin_parity(uint32_t x)
 {
     uint32_t p;
     volk_32u_popcnt(&p, x);
     return p & 1;
 }
-#endif /* __GNUC__ */
+#endif /* !defined(__GNUC__) && !defined(__clang__) */
 
 #define N 12