+++ python/__init__.py
@@ -19,7 +19,7 @@ stack, the AO-40 protocol used in the FUNcube satellites, and several
 ad-hoc protocols used in other satellites.
 '''
 
-__version__ = 'v3.6.0'
+__version__ = 'v3.7.0'
 __author__ = 'Daniel Estevez'
 __copyright__ = 'Copyright 2016-2020 Daniel Estevez'
 __email__ = 'daniel@destevez.net'
@@ -47,6 +47,8 @@ from .check_address import check_address
 from .check_ao40_uncoded_crc import check_ao40_uncoded_crc
 from .check_astrocast_crc import check_astrocast_crc
 from .check_cc11xx_crc import check_cc11xx_crc
+from .check_crc16_ccitt import check_crc16_ccitt
+from .check_crc16_ccitt_false import check_crc16_ccitt_false
 from .check_crc import check_crc
 from .check_eseo_crc import check_eseo_crc
 from .check_swiatowid_crc import check_swiatowid_crc