+++ python/check_crc16_ccitt.py
@@ -1,0 +1,57 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+# Copyright 2020 Daniel Estevez <daniel@destevez.net>
+#
+# This file is part of gr-satellites
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+
+import numpy
+from gnuradio import gr
+import pmt
+import struct
+
+from . import hdlc
+
+class check_crc16_ccitt(gr.basic_block):
+    """
+    docstring for block check_crc16_ccitt
+    """
+    def __init__(self, verbose):
+        gr.basic_block.__init__(self,
+            name="check_crc16_ccitt",
+            in_sig=[],
+            out_sig=[])
+
+        self.verbose = verbose
+        
+        self.message_port_register_in(pmt.intern('in'))
+        self.set_msg_handler(pmt.intern('in'), self.handle_msg)
+        self.message_port_register_out(pmt.intern('ok'))
+        self.message_port_register_out(pmt.intern('fail'))
+
+    def handle_msg(self, msg_pmt):
+        msg = pmt.cdr(msg_pmt)
+        if not pmt.is_u8vector(msg):
+            print("[ERROR] Received invalid message type. Expected u8vector")
+            return
+        packet = bytes(pmt.u8vector_elements(msg))
+
+        if len(packet) < 3:
+            return
+
+        packet_out = packet[:-2]
+        msg_out = pmt.cons(pmt.car(msg_pmt), pmt.init_u8vector(len(packet_out), packet_out))
+
+        crc = hdlc.crc_ccitt(packet_out)
+        packet_crc = struct.unpack("<H", packet[-2:])[0]
+        if crc == packet_crc:
+            if self.verbose:
+                print("CRC OK")
+            self.message_port_pub(pmt.intern('ok'), msg_out)
+        else:
+            if self.verbose:
+                print("CRC failed")
+            self.message_port_pub(pmt.intern('fail'), msg_out)